Split out the simple layer list from the basic sidebar too
